+++ manhattanDistance.hpp
@@ -1,6 +1,6 @@
 #ifndef MANHATTANDISTANCE_HPP
 #define MANHATTANDISTANCE_HPP
 
-unsigned int manhattanDist(class Point *, class Point *);
+double manhattanDist(class Point *, class Point *);
 
 #endif
